==> entrega_final/06_best_model.ipynb <==
# Best Model segun error absoluto de la predicción de 201912

# from google.colab import drive
import pandas as pd
import os


#  Definir rutas en Drive
BASE_INTERMEDIOS   = '/Users/indianaabeledo/Documents/Maestria/laboratorio_III/labo3-2025v/entrega_final/intermedios'
BASE_OUTPUTS   = '/Users/indianaabeledo/Documents/Maestria/laboratorio_III/labo3-2025v/entrega_final/output/'

os.makedirs(BASE_OUTPUTS, exist_ok=True)

# Carga de archivos desde Drive
print("🔄 Cargando archivos desde Drive...")
#pred_autogluon = pd.read_csv(os.path.join(BASE_OUTPUTS, "pred_modelo_autogluon.csv"))
pred_autogluon = pd.read_csv(os.path.join(BASE_OUTPUTS, "predicciones_modelo_Autogluonv2 (2).csv"))
pred_lgbm = pd.read_csv(os.path.join(BASE_OUTPUTS, "pred_modelo_lightgbm_best.csv"))
pred_RL= pd.read_csv(os.path.join(BASE_OUTPUTS, "pred_modelo_regresion_lineal.csv"))

error_autogluon = pd.read_csv(os.path.join(BASE_OUTPUTS, "error_201912_autogluon.csv"))
error_lgbm = pd.read_csv(os.path.join(BASE_OUTPUTS, "error_201912_lightgbm.csv"))
error_RL = pd.read_csv(os.path.join(BASE_OUTPUTS, "error_201912_regresion_lineal.csv"))

df_limpio= pd.read_csv(os.path.join(BASE_INTERMEDIOS, "df_limpio_product_id.csv"))

# Renames
pred_autogluon.rename(columns={'tn': 'tn_pred'}, inplace=True)
error_lgbm = error_lgbm.rename(columns={'target': 'tn_real'})  # renombrar 'tn' a 'tn_real'
# seleccionar solo las columnas deseadas
error_lgbm = error_lgbm[['product_id', 'tn_real', 'tn_pred', 'abs_error', 'rel_error']]

### Calculamos el promedio de la tn por product_id de los últimos 12 meses para 201912
Sera otra de las opciones de predicción

# Calcular promedio tn de los últimos 12 meses disponibles por product_id==201912
ultimos_meses = [201901, 201902, 201903, 201904, 201905, 201906, 201907, 201908, 201909, 201910, 201911,201912]
pred_prom_dic19_tn_12M = (
    df_limpio[df_limpio['periodo'].isin(ultimos_meses)]
      .groupby('product_id', as_index=False)
      .agg(avg=('tn','mean'))

)
pred_prom_dic19_tn_12M.head()

### Calculamos el promedio de la tn por product_id de los últimos 12 meses para 201910
Con el objetivo de calcular el error absoluto con la tn real

# Calcular promedio tn de los últimos 12 meses disponibles por product_id==201910
ultimos_meses = [201811, 201812,201901, 201902, 201903, 201904, 201905, 201906, 201907, 201908, 201909, 201910]
pred_prom_oct19_tn_12M = (
    df_limpio[df_limpio['periodo'].isin(ultimos_meses)]
      .groupby('product_id', as_index=False)
      .agg(avg=('tn','mean'))

)

pred_prom_oct19_tn_12M.head()

df_avg=df_limpio[['product_id','tn']][df_limpio['periodo']==201912]#aca lo cambie estaba 201910
df_avg.head()

### Calculamos el error absoluto entre el promedio de los últimos 12 meses para 201912 y la tn real


# Average: Merge entre tn y avg oct 2019
df_avg= pd.merge(df_avg,pred_prom_oct19_tn_12M, on='product_id', how='left')
df_avg.head()

# Calculo de error 201910
df_avg['abs_error_avg'] = (df_avg['tn'] - df_avg['avg']).abs()

df_avg.rename(columns={'avg': 'tn_pred_avg'}, inplace=True)
df_avg.drop(columns=['tn'], inplace=True)
df_avg.head()

### Unimos los errores absolutos de todos los modelos

# Merge por product_id y agregar sufijos para columnas comunes
error_merge = pd.merge(
    error_lgbm,
    error_autogluon,
    on='product_id',
    suffixes=('_lgbm', '_auto')
)
error_merge.head()

# 1) Renombrar tn_real_lgbm → tn_real
error_merge.rename(columns={'tn_real_lgbm':'tn_real'}, inplace=True)

# 2) Identificar todas las columnas a eliminar
cols_to_drop = [
    'tn_real_auto', 
] + [c for c in error_merge.columns 
     if 'rel_error' in c or 'pct_error' in c]

# 3) Eliminarlas en el mismo DataFrame
error_merge.drop(columns=cols_to_drop, inplace=True)

# 4) Comprobar resultado

error_merge.head()

error_merge_final = pd.merge(
    error_merge,
    error_RL,
    on='product_id')

error_merge_final.head()

# Partimos de tu df original
df = error_merge_final.copy()

# Lista de columnas de error
error_cols = ['abs_error_lgbm', 'abs_error_auto', 'abs_error_RL']

# 1) Calculamos para cada producto cuál columna tiene el valor mínimo
best_idx = df.set_index('product_id')[error_cols].idxmin(axis=1)

# 2) Extraemos el nombre del modelo a partir del sufijo de la columna
best_model = best_idx.str.replace('abs_error_', '')

# 3) Opcional: también recuperar el valor mínimo de error
min_error = df.set_index('product_id')[error_cols].min(axis=1)

# 4) Construimos el DataFrame resultante
df_best = pd.DataFrame({
    'product_id': best_model.index,
    'best_model': best_model.values,
    'min_abs_error': min_error.values
}).reset_index(drop=True)

print(df_best.head())

pred_lgbm.rename(columns={'tn_pred': 'lgbm'}, inplace=True)
pred_autogluon.rename(columns={'tn_pred': 'auto'}, inplace=True)
pred_RL.rename(columns={'pred': 'RL'}, inplace=True)

merge_pred= pd.merge(pred_autogluon, pred_lgbm, on='product_id')
merge_pred= pd.merge(merge_pred, pred_RL, on='product_id')
merge_pred.shape[0]

merge_pred.head()

merge_pred_final= pd.merge(merge_pred, df_best, on='product_id')
merge_pred_final.shape[0]

merge_pred_final.head()

# Partimos de tu merge_pred_final
df = merge_pred_final.copy()

# 1) Definimos el mapeo entre el valor de best_model y la columna de predicción
col_map = {
    'auto':    'auto',
    'lgbm':    'lgbm',
    'RL':      'RL'
}

# 2) Creamos el nuevo DF con product_id y la predicción del mejor modelo
df_best_pred = pd.DataFrame({
    'product_id': df['product_id'],
    'best_pred':  df.apply(lambda row: row[col_map[row['best_model']]], axis=1)
})

df_best_pred.head()

df_best_pred.to_csv(os.path.join(BASE_OUTPUTS, "best_model_product_id.csv"), index=False)

Agregamos Promedio al Merge de RL/Auto/LGBM

df_final_avg = pd.merge(
    error_merge_final,
    df_avg,
    on='product_id')
df_final_avg.rename(columns={'abs_error': 'abs_error_avg'}, inplace=True)
df_final_avg.head()

df_final_avg = df_final_avg.merge(
    pred_prom_dic19_tn_12M[['product_id','avg']],  # sólo product_id y avg
    on='product_id',
    how='left'
)
df_final_avg.head()

# Lista de columnas de error
error_cols = ['abs_error_lgbm', 'abs_error_auto', 'abs_error_RL', 'abs_error_avg']

# 1) Calculamos para cada producto cuál columna tiene el valor mínimo
best_idx_2 = df_final_avg.set_index('product_id')[error_cols].idxmin(axis=1)

# 2) Extraemos el nombre del modelo a partir del sufijo de la columna
best_model_2= best_idx_2.str.replace('abs_error_', '')

# 3) Opcional: también recuperar el valor mínimo de error
min_error_2 = df_final_avg.set_index('product_id')[error_cols].min(axis=1)

# 4) Construimos el DataFrame resultante
df_best_with_avg = pd.DataFrame({
    'product_id': best_model_2.index,
    'best_model': best_model_2.values,
    'min_abs_error': min_error_2.values
}).reset_index(drop=True)

print(df_best_with_avg.head())

df_best_with_avg['best_model'].value_counts(normalize=True)

pred_lgbm.rename(columns={'tn_pred': 'lgbm'}, inplace=True)
pred_autogluon.rename(columns={'tn_pred': 'auto'}, inplace=True)
pred_RL.rename(columns={'pred': 'RL'}, inplace=True)

merge_pred_2= pd.merge(pred_autogluon, pred_lgbm, on='product_id')
merge_pred_2= pd.merge(merge_pred_2, pred_RL, on='product_id')
merge_pred_2= pd.merge(merge_pred_2,pred_prom_dic19_tn_12M, on='product_id')
merge_pred_2.head()
merge_pred_final_2= pd.merge(merge_pred_2, df_best_with_avg, on='product_id')


merge_pred_final_2.head()

# 1) Definimos el mapeo entre el valor de best_model y la columna de predicción
col_map = {
    'auto':    'auto',
    'lgbm':    'lgbm',
    'RL':      'RL',
    'avg':    'avg'
}

# 2) Creamos el nuevo DF con product_id y la predicción del mejor modelo
df_best_pred_2 = pd.DataFrame({
    'product_id': merge_pred_final_2['product_id'],
    'best_pred':  merge_pred_final_2.apply(lambda row: row[col_map[row['best_model']]], axis=1)
})

df_best_pred_2.head()

df_best_pred_2.to_csv(os.path.join(BASE_OUTPUTS, "best_model_product_id_with_avg.csv"), index=False)

## PREDICCION CON PROMEDIO SIMPLE DE Autogluon /LGBM /RL / Avg 12 M

# Partimos de tu DataFrame original
prediccion_ponderada = merge_pred_final_2[['product_id','auto','lgbm','RL','avg']].copy()

# Calculamos pred = (auto + lgbm + RL + avg*0.98) / 4
prediccion_ponderada['pred'] = (
      prediccion_ponderada['auto']
    + prediccion_ponderada['lgbm']
    + prediccion_ponderada['RL']
    + prediccion_ponderada['avg'] * 1
) / 4

# 2) Eliminamos las columnas originales
prediccion_ponderada = prediccion_ponderada.drop(
    ['auto','lgbm','RL','avg'],
    axis=1
)

# Ahora prediccion_ponderada sólo tiene product_id y pred
print(prediccion_ponderada.head())

prediccion_ponderada.to_csv(os.path.join(BASE_OUTPUTS, "best_model_product_id_pond_final.csv"), index=False)

## PREDICCION CON PROMEDIO SIMPLE DE Autogluon /LGBM

# Partimos de tu DataFrame original
prediccion_ponderada_auto_lgbm = merge_pred_final_2[['product_id','auto','lgbm']].copy()

# 1) Creamos la columna 'pred' como la media de las 4 columnas
prediccion_ponderada_auto_lgbm['pred'] = prediccion_ponderada_auto_lgbm[
    ['auto','lgbm']
].mean(axis=1)

# 2) Eliminamos las columnas originales
prediccion_ponderada_auto_lgbm = prediccion_ponderada_auto_lgbm.drop(
    ['auto','lgbm'],
    axis=1
)

# Ahora prediccion_ponderada sólo tiene product_id y pred

print(prediccion_ponderada_auto_lgbm.head())
prediccion_ponderada_auto_lgbm.to_csv(os.path.join(BASE_OUTPUTS, "best_model_product_id_pond_auto_lgbm.csv"), index=False)

## PREDICCION CON PROMEDIO SIMPLE DE Autogluon /LGBM /RL 

# Partimos de tu DataFrame original
prediccion_ponderada_auto_lgbm_RL= merge_pred_final_2[['product_id','auto','lgbm','RL']].copy()

# 1) Creamos la columna 'pred' como la media de las 4 columnas
prediccion_ponderada_auto_lgbm_RL['pred'] = prediccion_ponderada_auto_lgbm_RL[['auto','lgbm','RL']
].mean(axis=1)

# 2) Eliminamos las columnas originales
prediccion_ponderada_auto_lgbm_RL = prediccion_ponderada_auto_lgbm_RL.drop(
    ['auto','lgbm','RL'],
    axis=1
)

# Ahora prediccion_ponderada sólo tiene product_id y pred

print(prediccion_ponderada_auto_lgbm_RL.head())
prediccion_ponderada_auto_lgbm_RL.to_csv(os.path.join(BASE_OUTPUTS, "best_model_product_id_pond_final_without_avg.csv"), index=False)

## PREDICCION MEDIANA DE LGBM /RL/AUTOGLUON

# Partimos de tu DataFrame original
prediccion_mediana = merge_pred_final_2[['product_id','auto','lgbm','RL']].copy()

# 1) Creamos la columna 'pred' como la mediana de las 3 columnas
prediccion_mediana['pred'] = prediccion_mediana[['auto','lgbm','RL']].median(axis=1)

# 2) Eliminamos las columnas originales
prediccion_mediana.drop(columns=['auto','lgbm','RL'], inplace=True)

print(prediccion_mediana.head())

# Ahora prediccion_ponderada sólo tiene product_id y pred

print(prediccion_mediana.head())
prediccion_mediana.to_csv(os.path.join(BASE_OUTPUTS, "best_model_product_id_mediana_final_without_avg.csv"), index=False)

## PREDICCION CON PROMEDIO SIMPLE DE LGBM /RL 

# Partimos de tu DataFrame original
prediccion_ponderada_lgbm_RL= merge_pred_final_2[['product_id','lgbm','RL']].copy()

# 1) Creamos la columna 'pred' como la media de las 4 columnas
prediccion_ponderada_lgbm_RL['pred'] = prediccion_ponderada_lgbm_RL[['lgbm','RL']
].mean(axis=1)

# 2) Eliminamos las columnas originales
prediccion_ponderada_lgbm_RL = prediccion_ponderada_lgbm_RL.drop(
    ['lgbm','RL'],
    axis=1
)

# Ahora prediccion_ponderada sólo tiene product_id y pred

print(prediccion_ponderada_lgbm_RL.head())
prediccion_ponderada_lgbm_RL.to_csv(os.path.join(BASE_OUTPUTS, "best_model_product_id_pond_lgbm_RL.csv"), index=False)

## Ensamble “mínima varianza” usando la matriz de covarianzas de los residuos

# Partimos de df_final_avg con columnas abs_error_* y predicciones
df_201910= df_final_avg.copy()
df_201910.drop(columns=['avg'], inplace=True)
df_201910.head()


#### Calcula los residuos de 201910 (predicción – valor real) para cada modelo:

# df201910: DataFrame con columnas ['product_id','pred_auto','pred_lgbm','pred_RL',,'tn_real']
res = df_201910[['tn_pred_auto','tn_pred_lgbm','tn_pred_RL']].values \
      - df_201910['tn_real'].values[:,None]   # shape (n_products,4)

### Mide la covarianza de esos residuos y calcula los pesos que minimizan la varianza del ensamblado:

import numpy as np

# 2.1) Matriz de covarianzas entre los residuos
Σ = np.cov(res, rowvar=False)       # 4×4

# 2.2) Pesos mínimos‐varianza: w = Σ⁻¹·1 / (1ᵀ·Σ⁻¹·1)
Σ_inv = np.linalg.pinv(Σ)
ones  = np.ones(Σ.shape[0])
w_mv  = Σ_inv.dot(ones) / (ones.dot(Σ_inv).dot(ones))

print("Pesos MV por modelo:", w_mv)

### Aplica esos pesos a tus predicciones de diciembre-2019:

#Armamos Predicciones 201912
df_201912=merge_pred_final_2.copy()
df_201912.drop(columns=['best_model'], inplace=True)
df_201912.drop(columns=['min_abs_error'], inplace=True)
df_201912.head()

# df_201912: DataFrame con ['product_id','pred_auto_12','pred_lgbm_12','pred_RL_12',]
X12 = df_201912[['auto','lgbm','RL']].values
df_201912['pred_mv'] = X12.dot(w_mv)

#Eliminamos las columnas originales
pred_var= df_201912.drop(
    ['auto','lgbm','RL','avg'],
    axis=1
)

# Generamos csv
print(pred_var.head())
pred_var.to_csv(os.path.join(BASE_OUTPUTS, "best_model_pred_var.csv"), index=False)

## Metamodelo de regresión lineal con intercepto (“stacking” ligero)

df_201910.head()

from sklearn.linear_model import LinearRegression

# 1) Preparamos X10, y10 de octubre-2019
X10 = df_201910[['tn_pred_auto','tn_pred_lgbm','tn_pred_RL']].values
y10 = df_201910['tn_real'].values

# 2) Entrenamos regresión CON intercepto
lr = LinearRegression(fit_intercept=True)
lr.fit(X10, y10)

print("Intercept:", lr.intercept_)
print("Pesos LR:", lr.coef_)

# 3) Aplicamos a diciembre-2019
X12 = df_201912[['auto','lgbm','RL',]].values
df_201912['pred_stack'] = lr.predict(X12)

df_201912.head()

# Eliminamos las columnas originales
pred_stack= df_201912.drop(
    ['auto','lgbm','RL','avg','pred_mv'],
    axis=1
)
# Generamos csv
print(pred_stack.head())
pred_stack.to_csv(os.path.join(BASE_OUTPUTS, "best_model_pred_stack.csv"), index=False)
